Move some files into separate modules
No functional changes; just moving things around.
