+++ src/curve/curve_multiplication.rs
