+++ src/curve/curve_summations.rs
